Address ledger review feedback
