+++ src/shared/ledger-links.ts
@@ -1,0 +1,3 @@
+/** The full-ledger view scoped to one listing's revenue and servicing costs. */
+export const listingLedgerHref = (listingId: number): string =>
+  `/admin/ledger?listing=${listingId}`;
